+++ packages/article-main-standard/src/styles/article-meta/index.android.js
@@ -1,9 +1,7 @@
 import { StyleSheet } from "react-native";
-import globalStyle from "../shared";
 import sharedStyles from "./shared";
 
 const styles = StyleSheet.create({
-  ...globalStyle,
   ...sharedStyles,
   articleMetaElement: {
     ...sharedStyles.articleMetaElement