+++ packages/article-main-standard/src/styles/article-meta/index.js
@@ -1,16 +1,8 @@
 import { StyleSheet } from "react-native";
-import globalStyle from "../shared";
 import sharedStyles from "./shared";
 
 const styles = StyleSheet.create({
-  ...globalStyle,
-  ...sharedStyles,
-  articleMetaElement: {
-    ...sharedStyles.articleMetaElement
-  },
-  byline: {
-    ...sharedStyles.byline
-  }
+  ...sharedStyles
 });
 
 export default styles;