+++ packages/article-skeleton/src/article-skeleton.web.js
@@ -17,6 +17,7 @@ import insertDropcapIntoAST from "./dropcap-util";
 import {
   BodyContainer,
   HeaderAdContainer,
+  HeaderContainer,
   MainContainer
 } from "./styles/responsive";
 import Head from "./head";
@@ -115,28 +116,30 @@ class ArticleSkeleton extends Component {
                   />
                 </HeaderAdContainer>
                 <MainContainer>
-                  <Header width={articleWidth} />
-                  {savingEnabled || sharingEnabled ? (
-                    <UserState state={UserState.loggedInOrShared}>
-                      <MessageContext.Consumer>
-                        {({ showMessage }) => (
-                          <StickySaveAndShareBar
-                            articleId={articleId}
-                            articleHeadline={headline}
-                            articleUrl={url}
-                            onCopyLink={() =>
-                              showMessage("Article link copied")
-                            }
-                            onSaveToMyArticles={() => {}}
-                            onShareOnEmail={() => {}}
-                            saveApi={saveApi}
-                            savingEnabled={savingEnabled}
-                            sharingEnabled={sharingEnabled}
-                          />
-                        )}
-                      </MessageContext.Consumer>
-                    </UserState>
-                  ) : null}
+                  <HeaderContainer>
+                    <Header width={articleWidth} />
+                    {savingEnabled || sharingEnabled ? (
+                      <UserState state={UserState.loggedInOrShared}>
+                        <MessageContext.Consumer>
+                          {({ showMessage }) => (
+                            <StickySaveAndShareBar
+                              articleId={articleId}
+                              articleHeadline={headline}
+                              articleUrl={url}
+                              onCopyLink={() =>
+                                showMessage("Article link copied")
+                              }
+                              onSaveToMyArticles={() => {}}
+                              onShareOnEmail={() => {}}
+                              saveApi={saveApi}
+                              savingEnabled={savingEnabled}
+                              sharingEnabled={sharingEnabled}
+                            />
+                          )}
+                        </MessageContext.Consumer>
+                      </UserState>
+                    ) : null}
+                  </HeaderContainer>
                   <BodyContainer>
                     <ArticleBody
                       content={newContent}
@@ -180,5 +183,8 @@ ArticleSkeleton.propTypes = {
 };
 ArticleSkeleton.defaultProps = articleSkeletonDefaultProps;
 
+export { KeylineItem, ArticleKeylineItem } from "./keylines";
+
 export { ArticleLink };
+
 export default articleTrackingContext(withTrackScrollDepth(ArticleSkeleton));