+++ packages/article-skeleton/src/styles/responsive.web.js
@@ -1,17 +1,15 @@
 import { View } from "react-native";
 import styled from "styled-components";
 import { breakpoints, colours, spacing } from "@times-components/styleguide";
-import { STICKY_CLASS_NAME } from "@times-components/sticky";
+import { ArticleKeylineItem } from "../keylines";
 
 export const MainContainer = styled(View)`
-  display: block;
   @media (min-width: ${breakpoints.wide}px) {
     padding-top: ${spacing(4)};
     margin: 0 auto;
   }
 `;
 
-/* --- HeaderAd --- */
 export const HeaderAdContainer = styled(View)`
   display: none;
 
@@ -25,62 +23,17 @@ export const HeaderAdContainer = styled(View)`
   }
 `;
 
+export const HeaderContainer = styled(View)`
+  display: flex;
+
+  /* stylelint-disable */
+  > ${ArticleKeylineItem} {
+    order: 3;
+  }
+  /* stylelint-enable */
+`;
+
 export const BodyContainer = styled(View)`
   display: block;
-  order: 3;
-`;
-
-/* --- SaveAndShareBar --- */
-export const SaveShareContainer = styled.div`
-  position: relative;
-  margin: 0 auto;
-  height: 60px;
-
-  @media (max-width: ${breakpoints.medium - 1}px) {
-    padding: 0 ${spacing(2)};
-  }
-
-  @media (min-width: ${breakpoints.medium}px) {
-    width: 80.8%;
-  }
-
-  @media (min-width: ${breakpoints.wide}px) {
-    width: 56.2%;
-  }
-
-  &:before,
-  &:after {
-    content: "";
-    height: 1px;
-    background-color: ${colours.functional.keyline};
-    position: absolute;
-    left: 0;
-    right: 0;
-  }
-
-  &:before {
-    top: 0;
-  }
-
-  &:after {
-    bottom: 0;
-  }
-`;
-
-export const OuterSaveShareContainer = styled.div`
-  margin: 0px auto ${spacing(6)};
-  background-color: ${colours.functional.white};
-
-  &.${STICKY_CLASS_NAME} {
-    box-shadow: 0 2px 8px 0 rgba(0, 0, 0, 0.3);
-
-    /* stylelint-disable */
-    ${SaveShareContainer} {
-      &:before,
-      &:after {
-        display: none;
-      }
-    }
-    /* stylelint-enable */
-  }
+  order: 4;
 `;