+++ packages/tracking/storybook-components/box.js
@@ -13,9 +13,9 @@ export const boxStyles = StyleSheet.create({
 });
 
 const Box = ({ children, color, id }) => (
-  <View id={id} style={[boxStyles.box, { backgroundColor: color }]}>
-    {children}
-  </View>
+  <div id={id}>
+    <View style={[boxStyles.box, { backgroundColor: color }]}>{children}</View>
+  </div>
 );
 Box.propTypes = {
   children: PropTypes.element,