+++ packages/article-main-standard/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.9.9](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.9.8...@times-components/article-main-standard@3.9.9) (2019-07-09)
+
+
+### Bug Fixes
+
+* lead asset position should be responsive on the main standard article template ([#2101](https://github.com/newsuk/times-components/issues/2101)) ([2f470d1](https://github.com/newsuk/times-components/commit/2f470d1))
+
+
+
+
+
+## [3.9.8](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.9.7...@times-components/article-main-standard@3.9.8) (2019-07-09)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+## [3.9.7](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.9.6...@times-components/article-main-standard@3.9.7) (2019-07-08)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+## [3.9.6](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.9.5...@times-components/article-main-standard@3.9.6) (2019-07-08)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
 ## [3.9.5](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.9.4...@times-components/article-main-standard@3.9.5) (2019-07-04)
 
 **Note:** Version bump only for package @times-components/article-main-standard