+++ packages/article/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.68](https://github.com/newsuk/times-components/compare/@times-components/article@7.2.67...@times-components/article@7.2.68) (2019-07-09)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.2.67](https://github.com/newsuk/times-components/compare/@times-components/article@7.2.66...@times-components/article@7.2.67) (2019-07-09)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.2.66](https://github.com/newsuk/times-components/compare/@times-components/article@7.2.65...@times-components/article@7.2.66) (2019-07-08)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.2.65](https://github.com/newsuk/times-components/compare/@times-components/article@7.2.64...@times-components/article@7.2.65) (2019-07-08)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.2.64](https://github.com/newsuk/times-components/compare/@times-components/article@7.2.63...@times-components/article@7.2.64) (2019-07-04)
 
 **Note:** Version bump only for package @times-components/article