+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.111](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.110...@times-components/pages@2.0.111) (2019-07-09)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.110](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.109...@times-components/pages@2.0.110) (2019-07-09)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.109](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.108...@times-components/pages@2.0.109) (2019-07-08)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.108](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.107...@times-components/pages@2.0.108) (2019-07-08)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [2.0.107](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.106...@times-components/pages@2.0.107) (2019-07-04)
 
 **Note:** Version bump only for package @times-components/pages